fix: install without :phoenix_live_view
Fix #8
